+++ rudof_lib/src/lib.rs
@@ -5,6 +5,8 @@
 pub mod rudof;
 pub mod rudof_config;
 pub mod rudof_error;
+pub mod shapes_graph_source;
 pub use rudof::*;
 pub use rudof_config::*;
 pub use rudof_error::*;
+pub use shapes_graph_source::*;