+++ rudof_lib/src/rudof.rs
@@ -1,10 +1,9 @@
-use crate::{RudofConfig, RudofError};
+use crate::{RudofConfig, RudofError, ShapesGraphSource};
 use iri_s::IriS;
 use prefixmap::PrefixMap;
 use shacl_ast::ast::Schema as ShaclSchema;
-use shacl_ast::compiled::schema::CompiledSchema as ShaclCompiledSchema;
 use shacl_ast::ShaclParser;
-use shacl_validation::shacl_processor::{EndpointValidation, GraphValidation, ShaclProcessor};
+use shacl_validation::shacl_processor::{GraphValidation, ShaclProcessor};
 use shacl_validation::store::graph::Graph;
 use shacl_validation::validation_report::report::ValidationReport;
 use shapemap::{query_shape_map::QueryShapeMap, ResultShapeMap};
@@ -34,7 +33,7 @@ pub struct Rudof {
     config: RudofConfig,
     rdf_data: RdfData,
     shex_schema: Option<ShExSchema>,
-    // shacl_schema: Option<ShaclSchema>, // TODO: Should we store a compiled schema to avoid compiling it for each validation request?
+    shacl_schema: Option<ShaclSchema>, // TODO: Should we store a compiled schema to avoid compiling it for each validation request?
     resolved_shex_schema: Option<SchemaWithoutImports>,
     shex_validator: Option<ShExValidator>,
     shapemap: Option<QueryShapeMap>,
@@ -45,7 +44,7 @@ impl Rudof {
         Rudof {
             config: config.clone(),
             shex_schema: None,
-            // shacl_schema: None,
+            shacl_schema: None,
             resolved_shex_schema: None,
             shex_validator: None,
             rdf_data: RdfData::new(),
@@ -53,37 +52,75 @@ impl Rudof {
         }
     }
 
-    /// Parse a SHACL schema from the current RDF data
-    pub fn get_shacl(&self) -> Result<ShaclSchema> {
+    /// Get the shapes graph schema from the current RDF data
+    pub fn get_shacl_from_data(&mut self) -> Result<()> {
         let schema = shacl_schema_from_data(self.rdf_data.clone())?;
-        Ok(schema)
+        self.shacl_schema = Some(schema);
+        Ok(())
     }
 
-    /*    /// Reads a `ShaclSchema`
-        /// It also updates the current Shacl processor with the new ShaclSchema
-        /// - `base` is used to resolve relative IRIs
-        /// - `format` indicates the Shacl format
-        pub fn shacl_schema_from_reader<R: io::Read>(
-            &self,
-            reader: R,
-            base: Option<&str>,
-            format: &ShaclFormat,
-            reader_mode: &ReaderMode,
-        ) -> Result<ShaclSchema> {
-            let format = match format {
-                ShaclFormat::Internal => Err(RudofError::InternalSHACLFormatNonReadable),
-                ShaclFormat::Turtle => Ok(RDFFormat::Turtle),
-                ShaclFormat::NTriples => Ok(RDFFormat::NTriples),
-                ShaclFormat::RDFXML => Ok(RDFFormat::RDFXML),
-                ShaclFormat::TriG => Ok(RDFFormat::TriG),
-                ShaclFormat::N3 => Ok(RDFFormat::N3),
-                ShaclFormat::NQuads => Ok(RDFFormat::NQuads),
-            }?;
-            let rdf_graph = parse_data(reader, &format, base, reader_mode)?;
-            let schema = shacl_schema_from_data(rdf_graph)?;
-            Ok(schema)
+    pub fn get_shacl(&self) -> Result<&ShaclSchema> {
+        if let Some(shacl_schema) = &self.shacl_schema {
+            Ok(&shacl_schema)
+        } else {
+            Err(RudofError::NoShaclSchema)
         }
-    */
+    }
+
+    pub fn get_shex(&self) -> Result<&ShExSchema> {
+        if let Some(shex_schema) = &self.shex_schema {
+            Ok(&shex_schema)
+        } else {
+            Err(RudofError::NoShaclSchema)
+        }
+    }
+
+    /// Resets the current validator
+    /// The action is necessary to start a fresh validation
+    pub fn reset_validation_results(&mut self) {
+        // TODO: We could add another operation to reset only the current validation results keeping the compiled schema
+        if let Some(ref mut validator) = &mut self.shex_validator {
+            validator.reset_result_map()
+        }
+    }
+
+    /// Resets the current validator
+    /// This operation removes the current shex_schema
+    pub fn reset_shex(&mut self) {
+        self.shex_schema = None;
+        self.shex_validator = None
+    }
+
+    /// Reads a SHACL schema from a reader
+    /// - `base` is used to resolve relative IRIs
+    /// - `format` indicates the Shacl format
+    pub fn read_shacl<R: io::Read>(
+        &mut self,
+        reader: R,
+        format: &ShaclFormat,
+        base: Option<&str>,
+        reader_mode: &ReaderMode,
+    ) -> Result<()> {
+        let format = match format {
+            ShaclFormat::Internal => Err(RudofError::InternalSHACLFormatNonReadable),
+            ShaclFormat::Turtle => Ok(RDFFormat::Turtle),
+            ShaclFormat::NTriples => Ok(RDFFormat::NTriples),
+            ShaclFormat::RDFXML => Ok(RDFFormat::RDFXML),
+            ShaclFormat::TriG => Ok(RDFFormat::TriG),
+            ShaclFormat::N3 => Ok(RDFFormat::N3),
+            ShaclFormat::NQuads => Ok(RDFFormat::NQuads),
+        }?;
+
+        let rdf_graph =
+            SRDFGraph::from_reader(reader, &format, base, reader_mode).map_err(|e| {
+                RudofError::ReadError {
+                    error: format!("{e}"),
+                }
+            })?;
+        let schema = shacl_schema_from_data(rdf_graph)?;
+        self.shacl_schema = Some(schema);
+        Ok(())
+    }
 
     /// Reads a `ShExSchema` and replaces the current one
     /// It also updates the current ShEx validator with the new ShExSchema
@@ -144,55 +181,46 @@ impl Rudof {
         Ok(())
     }
 
-    /* Get the current SHACL Schema
-    pub fn shacl_schema(&self) -> Option<&ShaclSchema> {
-        self.shacl_schema.as_ref()
-    } */
-
-    pub fn validate_shacl(&mut self, mode: ShaclValidationMode) -> Result<ValidationReport> {
-        let schema = shacl_schema_from_data(self.rdf_data.clone())?;
-        let schema_ast = Box::new(schema.clone());
-        if let Some(data) = self.rdf_data.graph() {
-            let validator = GraphValidation::from_graph(Graph::from_graph(data.to_owned()), mode);
-            let compiled_schema: ShaclCompiledSchema<SRDFGraph> = schema
-                .to_owned()
-                .try_into()
-                .map_err(|e| RudofError::SHACLCompilationError {
-                    error: format!("{e}"),
-                    schema: schema_ast.clone(),
-                })?;
-            let result = ShaclProcessor::validate(&validator, &compiled_schema).map_err(|e| {
-                RudofError::SHACLValidationError {
-                    error: format!("{e}"),
-                    schema: schema_ast,
-                }
-            })?;
-            Ok(result)
-        } else if let Some(endpoint) = self.rdf_data.first_endpoint() {
-            let validator =
-                EndpointValidation::from_sparql(endpoint.to_owned(), mode).map_err(|e| {
-                    RudofError::SHACLEndpointValidationCreation {
-                        endpoint: endpoint.clone(),
+    /// Validate RDF data using SHACL
+    ///
+    /// mode: Indicates whether to use SPARQL or native Rust implementation
+    /// shapes_graph_source: Indicates the source of the shapes graph: either from the current data, or from the current SHACL schema
+    /// If there is no current SHACL schema, it tries to get it from the current RDF data
+    pub fn validate_shacl(
+        &mut self,
+        mode: ShaclValidationMode,
+        shapes_graph_source: ShapesGraphSource,
+    ) -> Result<ValidationReport> {
+        let (compiled_schema, shacl_schema) = match shapes_graph_source {
+            ShapesGraphSource::CurrentSchema if self.shacl_schema.is_some() => {
+                let ast_schema = self.shacl_schema.as_ref().unwrap();
+                let compiled_schema = ast_schema.clone().to_owned().try_into().map_err(|e| {
+                    RudofError::SHACLCompilationError {
                         error: format!("{e}"),
+                        schema: Box::new(ast_schema.clone()),
                     }
                 })?;
-            let compiled_schema: ShaclCompiledSchema<SRDFSparql> = schema
-                .to_owned()
-                .try_into()
-                .map_err(|e| RudofError::SHACLCompilationError {
-                    error: format!("{e}"),
-                    schema: schema_ast.clone(),
+                Ok((compiled_schema, ast_schema.clone()))
+            }
+            _ => {
+                let ast_schema = shacl_schema_from_data(self.rdf_data.clone())?;
+                let compiled_schema = ast_schema.to_owned().try_into().map_err(|e| {
+                    RudofError::SHACLCompilationError {
+                        error: format!("{e}"),
+                        schema: Box::new(ast_schema.clone()),
+                    }
                 })?;
-            let result = ShaclProcessor::validate(&validator, &compiled_schema).map_err(|e| {
-                RudofError::SHACLValidationError {
-                    error: format!("{e}"),
-                    schema: schema_ast,
-                }
-            })?;
-            Ok(result)
-        } else {
-            Err(RudofError::NoGraphNoFirstEndpoint)
-        }
+                Ok((compiled_schema, ast_schema))
+            }
+        }?;
+        let validator = GraphValidation::from_graph(Graph::from_data(self.rdf_data.clone()), mode);
+        let result = ShaclProcessor::validate(&validator, &compiled_schema).map_err(|e| {
+            RudofError::SHACLValidationError {
+                error: format!("{e}"),
+                schema: Box::new(shacl_schema),
+            }
+        })?;
+        Ok(result)
     }
 
     pub fn validate_shex(&mut self) -> Result<ResultShapeMap> {
@@ -373,6 +401,8 @@ fn shacl_schema_from_data<RDF: FocusRDF + Debug>(rdf_data: RDF) -> Result<ShaclS
 #[cfg(test)]
 mod tests {
     use iri_s::iri;
+    use shacl_ast::ShaclFormat;
+    use shacl_validation::shacl_processor::ShaclValidationMode;
     use shapemap::ShapeMapFormat;
     use shex_ast::{compiled::shape_label::ShapeLabel, Node};
     use shex_validation::ShExFormat;
@@ -433,5 +463,95 @@ mod tests {
         let node = Node::iri(iri!("http://example/x"));
         let shape = ShapeLabel::iri(iri!("http://example/S"));
         assert!(result.get_info(&node, &shape).unwrap().is_non_conformant(),)
+    }
+
+    #[test]
+    fn test_shacl_validation_ok() {
+        let data = r#"prefix : <http://example.org/> 
+        :x :p 23 .
+        "#;
+        let shacl = r#"prefix :       <http://example.org/> 
+            prefix sh:     <http://www.w3.org/ns/shacl#> 
+            prefix xsd:    <http://www.w3.org/2001/XMLSchema#> 
+
+            :S a sh:NodeShape; sh:closed true ;
+              sh:targetNode :x ;
+            sh:property [                  
+                sh:path     :p ; 
+                sh:minCount 1; 
+                sh:maxCount 1;
+                sh:datatype xsd:integer ;
+            ] .
+             "#;
+        let mut rudof = Rudof::new(&RudofConfig::default());
+        rudof
+            .merge_data_from_reader(
+                data.as_bytes(),
+                &srdf::RDFFormat::Turtle,
+                None,
+                &srdf::ReaderMode::Strict,
+            )
+            .unwrap();
+
+        rudof
+            .read_shacl(
+                shacl.as_bytes(),
+                &ShaclFormat::Turtle,
+                None,
+                &srdf::ReaderMode::Lax,
+            )
+            .unwrap();
+        let result = rudof
+            .validate_shacl(
+                ShaclValidationMode::Native,
+                crate::ShapesGraphSource::CurrentSchema,
+            )
+            .unwrap();
+        assert!(result.results().is_empty())
+    }
+
+    #[test]
+    fn test_shacl_validation_ko() {
+        let data = r#"prefix : <http://example.org/> 
+        :x :other 23 .
+        "#;
+        let shacl = r#"prefix :       <http://example.org/> 
+            prefix sh:     <http://www.w3.org/ns/shacl#> 
+            prefix xsd:    <http://www.w3.org/2001/XMLSchema#> 
+
+            :S a sh:NodeShape; sh:closed true ;
+             sh:targetNode :x ; 
+            sh:property [                  
+                sh:path     :p ; 
+                sh:minCount 1; 
+                sh:maxCount 1;
+                sh:datatype xsd:integer ;
+            ] .
+             "#;
+        let mut rudof = Rudof::new(&RudofConfig::default());
+        rudof
+            .merge_data_from_reader(
+                data.as_bytes(),
+                &srdf::RDFFormat::Turtle,
+                None,
+                &srdf::ReaderMode::Strict,
+            )
+            .unwrap();
+
+        rudof
+            .read_shacl(
+                shacl.as_bytes(),
+                &ShaclFormat::Turtle,
+                None,
+                &srdf::ReaderMode::Lax,
+            )
+            .unwrap();
+        let result = rudof
+            .validate_shacl(
+                ShaclValidationMode::Native,
+                crate::ShapesGraphSource::CurrentSchema,
+            )
+            .unwrap();
+        assert!(result.results().is_empty())
     }
 }