+++ shacl_validation/src/shape.rs
@@ -76,7 +76,7 @@ impl<S: SRDFBasic> FocusNodesOps<S> for CompiledShape<S> {
     fn focus_nodes(&self, store: &S, runner: &dyn Engine<S>) -> FocusNodes<S> {
         runner
             .focus_nodes(store, self, self.targets())
-            .expect("Failed to retrieve focus nodes") // TODO: expect?
+            .expect("Failed to retrieve focus nodes")
     }
 }