+++ shacl_validation/src/store/graph.rs
@@ -1,5 +1,6 @@
 use std::path::Path;
 
+use sparql_service::RdfData;
 use srdf::{RDFFormat, ReaderMode, SRDFGraph};
 
 use crate::validate_error::ValidateError;
@@ -7,12 +8,17 @@ use crate::validate_error::ValidateError;
 use super::Store;
 
 pub struct Graph {
-    store: SRDFGraph,
+    store: RdfData,
 }
 
 impl Graph {
-    // TODO: I would change this to from_path
-    pub fn new(
+    pub fn new() -> Graph {
+        Graph {
+            store: RdfData::new(),
+        }
+    }
+
+    pub fn from_path(
         path: &Path,
         rdf_format: RDFFormat,
         base: Option<&str>,
@@ -23,18 +29,26 @@ impl Graph {
             base,
             &ReaderMode::default(), // TODO: this should be revisited
         ) {
-            Ok(store) => Ok(Self { store }),
+            Ok(store) => Ok(Self {
+                store: RdfData::from_graph(store)?,
+            }),
             Err(error) => Err(ValidateError::Graph(error)),
         }
     }
 
-    pub fn from_graph(graph: SRDFGraph) -> Graph {
-        Graph { store: graph }
+    pub fn from_graph(graph: SRDFGraph) -> Result<Graph, ValidateError> {
+        Ok(Graph {
+            store: RdfData::from_graph(graph)?,
+        })
+    }
+
+    pub fn from_data(data: RdfData) -> Graph {
+        Graph { store: data }
     }
 }
 
-impl Store<SRDFGraph> for Graph {
-    fn store(&self) -> &SRDFGraph {
+impl Store<RdfData> for Graph {
+    fn store(&self) -> &RdfData {
         &self.store
     }
 }