+++ shacl_validation/src/validate_error.rs
@@ -1,6 +1,7 @@
 // use oxiri::IriParseError;
 use shacl_ast::compiled::compiled_shacl_error::CompiledShaclError;
 use shacl_ast::shacl_parser_error::ShaclParserError;
+use sparql_service::RdfDataError;
 use srdf::RDFParseError;
 use srdf::SRDFGraphError;
 use thiserror::Error;
@@ -41,6 +42,8 @@ pub enum ValidateError {
     SrdfHelper(#[from] SRDFError),
     #[error("Error during the compilation of the Schema, {}", ._0)] // TODO: move to store
     CompiledShacl(#[from] CompiledShaclError),
-    #[error("Not yet implemented")]
-    NotImplemented,
+    #[error("Not yet implemented: {msg}")]
+    NotImplemented { msg: String },
+    #[error(transparent)]
+    RdfDataError(#[from] RdfDataError),
 }