+++ sparql_service/src/query_processor.rs
@@ -1,5 +1,5 @@
 use prefixmap::PrefixMap;
-use srdf::{QuerySolutionIter, SRDF};
+use srdf::{QuerySolutions, SRDF};
 
 use crate::RdfData;
 
@@ -19,7 +19,7 @@ impl QueryProcessor {
         Some(self.rdf_data.prefixmap_in_memory())
     }
 
-    pub fn query_select<S: SRDF>(_str: &str) -> QuerySolutionIter<S> {
+    pub fn query_select<S: SRDF>(_str: &str) -> QuerySolutions<S> {
         todo!()
     }
 }